requirements.txt
# -------------------------
# Core runtime
# -------------------------
python-dotenv>=1.0,<2.0
pydantic>=2.6,<3.0
typing-extensions>=4.10,<5.0
packaging>=23,<27
requests>=2.31,<3.0

# -------------------------
# UI
# -------------------------
streamlit>=1.31,<2.0
pandas>=2.0,<3.0
numpy>=1.26,<3.0

# -------------------------
# LangChain / LangGraph (v1-style, no meta-package)
# -------------------------
langchain-core>=0.2.30,<0.4
langchain-community>=0.2.16,<0.4
langchain-openai>=0.1.17,<0.3
langchain-text-splitters>=0.2.2,<0.4

langgraph>=1.0,<1.1

# -------------------------
# OpenAI + tokens
# -------------------------
openai>=1.12,<2.0
tiktoken>=0.6,<1.0

# -------------------------
# RAG / storage
# -------------------------
faiss-cpu>=1.8,<2.0
diskcache>=5.6,<6.0
sentence-transformers>=2.6,<3.0

# -------------------------
# Reporting
# -------------------------
reportlab>=4.0,<5.0

# -------------------------
# CAD ingestion (MVP, optional)
# -------------------------
# STEP parsing + bbox/face proxies for bin derivation. App runs without it (manual bins only).
# OCP = OpenCASCADE Python bindings (STEPControl_Reader, Bnd_Box, etc.); do not use pythonocc-core.
cadquery-ocp>=7.7,<8.0
cadquery>=2.4,<3.0
